+++ src/Commands/CreateMigrationsTable.php
@@ -5,18 +5,24 @@ declare(strict_types=1);
 namespace Celema\Quma\Commands;
 
 use Celema\Console\Args;
+use Celema\Console\Command;
+use Celema\Console\Io;
+use Celema\Quma\Connection;
+use Celema\Quma\Environment;
 use Celema\Quma\Migrations\MetadataTable;
-use Override;
 
-final class CreateMigrationsTable extends Command
+#[Command('db:create-migrations-table', 'Creates a migrations table', group: 'Database')]
+final class CreateMigrationsTable
 {
-	protected string $name = 'create-migrations-table';
-	protected string $group = 'Database';
-	protected string $prefix = 'db';
-	protected string $description = 'Creates a migrations table';
+	private readonly Environment $env;
 
-	#[Override]
-	public function run(Args $args): int
+	/** @param array<non-empty-string, Connection>|Connection $conn */
+	public function __construct(array|Connection $conn, array $options = [])
+	{
+		$this->env = new Environment($conn, $options);
+	}
+
+	public function __invoke(Args $args, Io $io): int
 	{
 		$env = $this->env;